Refactor sync metrics tests into modular helpers
